LPS-50752 create central upgrade process
Create a stable component to represent the latest upgrade process. This
way, other components that depend on the process to be executed don't
have to depend on a particular version.
